+++ db/migrate/20191205113156_remove_description_index_from_service.rb
@@ -1,0 +1,5 @@
+class RemoveDescriptionIndexFromService < ActiveRecord::Migration[5.2]
+  def change
+    remove_index :services, :description
+  end
+end